requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
joblib==1.0.1
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sqlparse==0.4.1
threadpoolctl==2.2.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
